pyproject.toml
[tool.poetry]
name = "ed-education-certificate-generator"
version = "0.1.0"
description = ""
authors = ["lchansky <lchansky@mail.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
requests = "^2.31.0"
pytelegrambotapi = "^4.12.0"
gunicorn = "^20.1.0"
openpyxl = "^3.1.2"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
